feat: moved some useful tools to common
