+++ src/core/client/embed/utils/index.ts
@@ -1,6 +1,4 @@
 export { default as buildURL } from "./buildURL";
-export { default as ensureEndSlash } from "./ensureEndSlash";
-export { default as ensureNoEndSlash } from "./ensureNoEndSlash";
 export { default as startsWith } from "./startsWith";
 export { default as prefixStorage } from "./prefixStorage";
 export { default as parseHashQuery } from "./parseHashQuery";
