+++ src/core/common/utils/index.ts
@@ -3,3 +3,5 @@ export { default as animationFrame } from "./animationFrame";
 export { default as pascalCase } from "./pascalCase";
 export { default as oncePerFrame } from "./oncePerFrame";
 export { default as isBeforeDate } from "./isBeforeDate";
+export { default as ensureEndSlash } from "./ensureEndSlash";
+export { default as ensureNoEndSlash } from "./ensureNoEndSlash";
